+++ app/Http/Controllers/Admin/DishController.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
-use App\Models\Type;
-use App\Http\Requests\StoreTypeRequest;
-use App\Http\Requests\UpdateTypeRequest;
+use App\Models\Dish;
+use App\Http\Requests\Admin\Dish\StoreDishRequest;
+use App\Http\Requests\Admin\Dish\UpdateDishRequest;
+use App\Http\Controllers\Controller;
 
-class TypeController extends Controller
+class DishController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -27,7 +28,7 @@ class TypeController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreTypeRequest $request)
+    public function store(StoreDishRequest $request)
     {
         //
     }
@@ -35,7 +36,7 @@ class TypeController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Type $type)
+    public function show(Dish $dish)
     {
         //
     }
@@ -43,7 +44,7 @@ class TypeController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Type $type)
+    public function edit(Dish $dish)
     {
         //
     }
@@ -51,7 +52,7 @@ class TypeController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateTypeRequest $request, Type $type)
+    public function update(UpdateDishRequest $request, Dish $dish)
     {
         //
     }
@@ -59,7 +60,7 @@ class TypeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Type $type)
+    public function destroy(Dish $dish)
     {
         //
     }
