+++ app/Http/Requests/Admin/Dish/UpdateDishRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Admin\Dish;
 
 use Illuminate\Foundation\Http\FormRequest;
