+++ app/Http/Requests/Admin/Type/StoreTypeRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Admin\Type;
 
 use Illuminate\Foundation\Http\FormRequest;
